Refactor package and object
