+++ src/main/scala/uk/gov/nationalarchives/tdr/transfer/service/api/TransferServiceServer.scala
@@ -17,7 +17,7 @@ import sttp.tapir.swagger.bundle.SwaggerInterpreter
 import uk.gov.nationalarchives.tdr.keycloak.TdrKeycloakDeployment
 import uk.gov.nationalarchives.tdr.transfer.service.ApplicationConfig
 import uk.gov.nationalarchives.tdr.transfer.service.api.controllers.LoadController
-import uk.gov.nationalarchives.tdr.transfer.service.api.cors.CustomCORSInterceptor
+import uk.gov.nationalarchives.tdr.transfer.service.api.interceptors.CustomInterceptors
 
 object TransferServiceServer extends IOApp {
   implicit def logger: SelfAwareStructuredLogger[IO] = Slf4jLogger.getLogger[IO]
@@ -46,7 +46,7 @@ object TransferServiceServer extends IOApp {
 
   private val customServerOptions: Http4sServerOptions[IO] = Http4sServerOptions
     .customiseInterceptors[IO]
-    .corsInterceptor(CustomCORSInterceptor.apply())
+    .corsInterceptor(CustomInterceptors.customCorsInterceptor)
     .options
 
   private val allRoutes =
